Fix CI

+++ code/non_planar_slicing_deformation/state/ThreeAxisDeformerState.py
@@ -1,7 +1,6 @@
 from dataclasses import dataclass
 
 import numpy as np
-from typing_extensions import Callable
 
 from non_planar_slicing_deformation.state.DeformerState import DeformerState
 
@@ -11,7 +10,7 @@ class ThreeAxisDeformerState(DeformerState):
     """
     The state for :class:`ThreeAxisDeformer` and :class:`ThreeAxisUndeformer`
     """
-    firstOrder: float
-    secondOrder: float
-    translationX: float
-    translationY: float
+    firstOrder: np.float64
+    secondOrder: np.float64
+    translationX: np.float64
+    translationY: np.float64
